requirements.txt
neat-python==0.92
numpy==1.14.3
gym==0.10.5
click==6.7
torch==0.4.0
